Yummly README.md
Once you have collected this information, you should then submit it to the relevant party. This might be via an in-app feature, a support email, a bug-tracking system, or another mechanism provided by Yummly or the developers responsible for the component in question.

In case Yummly is associated with a third-party application or service, you should report the bug to the developers or the support team of that particular service.

Remember to check the official support channels or the help section in the app or on the Yummly website for specific instructions or contact details for reporting bugs or issues.